Extracted all classes to their own dedicated file.
